// Front/src/app/models/ciudadano.model.ts
export interface Ciudadano {
  idCiudadano: number;
  nombre: string;
  apellidos: string;
  fechaNacimiento: string;
  calle: string;
  numero: number;
  localidad: string;
  codigoPostal: number;
  telefono: string;
  email: string;
  profesion: string;
  curriculumURL?: string;
  validado: boolean;
 
    idUsuario: number;
    nombreUsuario: string;
    perfil: string;
 
}